adding self-referencing associations to making sub Category and update category and product controller

+++ src/main/java/com/example/producttestapi/repos/UserRepo.java
@@ -3,9 +3,9 @@ package com.example.producttestapi.repos;
 import com.example.producttestapi.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User, Long> {
+    public interface UserRepo extends JpaRepository<User, Integer> {
     User findByEmail(String  email);
-    User getUserById(Long id);
+    User getUserById(int id);
     User findAllByEmail(String email);
-    void deleteById(Long id);
+    void deleteById(int id);
 }
