+++ src/main/java/com/example/producttestapi/service/VoucherService.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 public interface VoucherService {
     Voucher createVoucher(Voucher voucher);
     Voucher findVoucherByCode(String code);
-    void deleteVoucher(Long id);
+    void deleteVoucher(int id);
     void applyVoucherDiscount(Product product);
 }
